+++ impeller/renderer/compute_subgroup_unittests.cc
@@ -11,6 +11,7 @@
 #include "gmock/gmock.h"
 #include "impeller/base/strings.h"
 #include "impeller/core/formats.h"
+#include "impeller/display_list/skia_conversions.h"
 #include "impeller/entity/contents/content_context.h"
 #include "impeller/fixtures/cubic_to_quads.comp.h"
 #include "impeller/fixtures/golden_paths.h"
@@ -35,87 +36,6 @@
 
 namespace impeller {
 namespace testing {
-
-static Point ToPoint(const SkPoint& point) {
-  return Point::MakeXY(point.fX, point.fY);
-}
-
-static Path ToPath(const SkPath& path) {
-  auto iterator = SkPath::Iter(path, false);
-
-  struct PathData {
-    union {
-      SkPoint points[4];
-    };
-  };
-
-  PathBuilder builder;
-  PathData data;
-  auto verb = SkPath::Verb::kDone_Verb;
-  do {
-    verb = iterator.next(data.points);
-    switch (verb) {
-      case SkPath::kMove_Verb:
-        builder.MoveTo(ToPoint(data.points[0]));
-        break;
-      case SkPath::kLine_Verb:
-        builder.LineTo(ToPoint(data.points[1]));
-        break;
-      case SkPath::kQuad_Verb:
-        builder.QuadraticCurveTo(ToPoint(data.points[1]),
-                                 ToPoint(data.points[2]));
-        break;
-      case SkPath::kConic_Verb: {
-        constexpr auto kPow2 = 1;  // Only works for sweeps up to 90 degrees.
-        constexpr auto kQuadCount = 1 + (2 * (1 << kPow2));
-        SkPoint points[kQuadCount];
-        const auto curve_count =
-            SkPath::ConvertConicToQuads(data.points[0],          //
-                                        data.points[1],          //
-                                        data.points[2],          //
-                                        iterator.conicWeight(),  //
-                                        points,                  //
-                                        kPow2                    //
-            );
-
-        for (int curve_index = 0, point_index = 0;  //
-             curve_index < curve_count;             //
-             curve_index++, point_index += 2        //
-        ) {
-          builder.QuadraticCurveTo(ToPoint(points[point_index + 1]),
-                                   ToPoint(points[point_index + 2]));
-        }
-      } break;
-      case SkPath::kCubic_Verb:
-        builder.CubicCurveTo(ToPoint(data.points[1]), ToPoint(data.points[2]),
-                             ToPoint(data.points[3]));
-        break;
-      case SkPath::kClose_Verb:
-        builder.Close();
-        break;
-      case SkPath::kDone_Verb:
-        break;
-    }
-  } while (verb != SkPath::Verb::kDone_Verb);
-
-  FillType fill_type;
-  switch (path.getFillType()) {
-    case SkPathFillType::kWinding:
-      fill_type = FillType::kNonZero;
-      break;
-    case SkPathFillType::kEvenOdd:
-      fill_type = FillType::kOdd;
-      break;
-    case SkPathFillType::kInverseWinding:
-    case SkPathFillType::kInverseEvenOdd:
-      FML_DCHECK(false);
-      // Flutter doesn't expose these path fill types. These are only visible
-      // via the receiver interface. We should never get here.
-      fill_type = FillType::kNonZero;
-      break;
-  }
-  return builder.TakePath(fill_type);
-}
 
 using ComputeSubgroupTest = ComputePlaygroundTest;
 INSTANTIATE_COMPUTE_SUITE(ComputeSubgroupTest);
@@ -152,7 +72,7 @@ TEST_P(ComputeSubgroupTest, PathPlayground) {
 
     SkPath sk_path;
     if (SkParsePath::FromSVGString(svg_path_data, &sk_path)) {
-      auto path = ToPath(sk_path);
+      auto path = skia_conversions::ToPath(sk_path);
       auto status =
           ComputeTessellator{}
               .SetStrokeWidth(stroke_width)
@@ -245,6 +165,8 @@ TEST_P(ComputeSubgroupTest, PathPlayground) {
 }
 
 TEST_P(ComputeSubgroupTest, LargePath) {
+  // The path in here is large enough to highlight issues around exceeding
+  // subgroup size.
   using SS = StrokeComputeShader;
 
   auto context = GetContext();
